feat: Try to download the file if the OPTIONS fail
We are seeing too many 500 in the `file_info` method, that uses the
`OPTION` http method, with this commit we catch this error, log an error
and download the file anyway. If the file already exists, we would
replace it, but since we are saving with the hash, it's improbable that
we will override a valid file.

This commit also changes the prefix of the files from the remote id to
the document to provide a better file name.

Signed-off-by: Arturo Volpe <arturovolpe@gmail.com>