fix: RelationGraph
The react-sigma doesn't support a change in it's graph parameter, so we
need to pass the loaded data and not a empty array and later the full
data.

In a previous change, the data was changed from a simple array to a
Async, but the condition to load the graph was the same, and a async is
never undefined so the graph was rendered without data.

Signed-off-by: Arturo Volpe <arturovolpe@gmail.com>