feat: Use the new dynamic query on Affidavit table
The affidavit table was very slow because the previous version fetched
the full list of declarations, this new version only fetch a page at any
given moment.

A generic hook is implemented that knows how to talk with the api and
also knows how to talk with the antd table.

Signed-off-by: Arturo Volpe <arturovolpe@gmail.com>